cleanup, do we still have hot reloading on api routes?
