--- server/src/api/next/index.ts
@@ -1,6 +1,0 @@
-import { pluginFromRoutes } from "server/api/next/helpers"
-import simpleOpen from "server/api/next/open"
-
-const ApiPlugin = pluginFromRoutes([simpleOpen])
-
-export default ApiPlugin
